requirements.txt
# Core Dependencies
langgraph==0.2.28
langchain==0.3.1
langchain-openai==0.2.1
langchain-core==0.3.6

# OpenAI - Fixed version compatibility
openai==1.40.0
httpx==0.27.2  # Pin httpx to avoid proxies parameter issue with langchain-openai

# Observability
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation==0.48b0
python-json-logger==2.0.7

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
aiohttp==3.10.5

# Development
pytest==8.3.3
pytest-asyncio==0.24.0